Crate nix_bindings_util_sys

Raw bindings to Nix C API

This crate contains automatically generated bindings from the Nix C headers. The bindings are generated by bindgen and include C-style naming conventions and documentation comments that don’t always conform to Rust standards.

Normally you don’t have to use this crate directly. Instead use nix-util.

Structs

c_context

Constants

err_NIX_ERR_KEY

@brief A key/index access error occurred in C API functions.

err_NIX_ERR_NIX_ERROR

@brief A generic Nix error occurred.

err_NIX_ERR_OVERFLOW

@brief An overflow error occurred.

err_NIX_ERR_UNKNOWN

@brief An unknown error occurred.

err_NIX_OK

@brief No error occurred.

verbosity_NIX_LVL_CHATTY

verbosity_NIX_LVL_DEBUG

verbosity_NIX_LVL_ERROR

verbosity_NIX_LVL_INFO

verbosity_NIX_LVL_NOTICE

verbosity_NIX_LVL_TALKATIVE

verbosity_NIX_LVL_VOMIT

verbosity_NIX_LVL_WARN

Functions

c_context_create⚠

@brief Allocate a new nix_c_context. @throws std::bad_alloc @return allocated nix_c_context, owned by the caller. Free using nix_c_context_free.

c_context_free⚠

@brief Free a nix_c_context. Does not fail. @param[out] context The context to free, mandatory.

clear_err⚠

@brief Clear the error message from a nix context.

err_code⚠

@brief Retrieves the most recent error code from a nix_c_context

err_info_msg⚠

@brief Retrieves the error message from errorInfo in a context.

err_msg⚠

@addtogroup errors @{ / /** @brief Retrieves the most recent error message from a context.

err_name⚠

@brief Retrieves the error name from a context.

libutil_init⚠

@brief Initializes nix_libutil and its dependencies.

set_err_msg⚠

@brief Set an error message on a nix context.

set_verbosity⚠

@brief Sets the verbosity level

setting_get⚠

@defgroup settings Nix configuration settings @{ / /** @brief Retrieves a setting from the nix global configuration.

setting_set⚠

@brief Sets a setting in the nix global configuration.

version_get⚠

@brief Retrieves the nix library version.

Type Aliases

err

@brief Type for error codes in the Nix system

get_string_callback

@brief Called to get the value of a string owned by Nix.

verbosity

@brief Verbosity level