nix_bindings_util_sys

Function err_info_msg

pub unsafe extern "C" fn err_info_msg(
    context: *mut c_context,
    read_context: *const c_context,
    callback: get_string_callback,
    user_data: *mut c_void,
) -> err

@brief Retrieves the error message from errorInfo in a context.

Used to inspect nix Error messages.

@pre This function should only be called after a previous nix function has returned a NIX_ERR_NIX_ERROR

@param[out] context optional, the context to store errors in if this function fails @param[in] read_context the context to retrieve the error message from. @param[in] callback Called with the error message. @param[in] user_data optional, arbitrary data, passed to the callback when it’s called. @see nix_get_string_callback @return NIX_OK if there were no errors, an error code otherwise.