nix_bindings_util_sys

Function err_msg

pub unsafe extern "C" fn err_msg(
    context: *mut c_context,
    ctx: *const c_context,
    n: *mut c_uint,
) -> *const c_char

@addtogroup errors @{ / /** @brief Retrieves the most recent error message from a context.

@pre This function should only be called after a previous nix function has returned an error.

@param[out] context optional, the context to store errors in if this function fails @param[in] ctx the context to retrieve the error message from @param[out] n optional: a pointer to an unsigned int that is set to the length of the error. @return nullptr if no error message was ever set, a borrowed pointer to the error message otherwise, which is valid until the next call to a Nix function, or until the context is destroyed.